requirements-dev.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==7.2.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
